Implement Drag'n'drop. Add more error checking.
This commit implements drag and drop and the relevant checking for
making sure everything works well. In particular, only local files are
permitted (the player can't play anything directly from the internet)
and of those, gme_open_file will make sure to error out if any non-game-music
ones are dropped.
This commit also indirectly fixes open_file so that it shows an error
message if the user tries to open a non-music file from the file dialog.