Integrate tl::expected, fix file error handling.
This commit updates io.hpp to use an error handling facility that should
be introduced in C++23. However, since it is currently 2022, I am using
an alternative library instead.
The changes to the other files should fix any problem regarding opening
files and displaying the right errors to the user.
a.playlist was modified to be able to test the error handling.