Huge refactoring.
This commit does a big refactoring on GUI code.

Most of the changes are based on executing player actions in a
consistent way, i.e., in a widget callback we shall always only call
player methods and never do any update to other widgets (unless
absolutely necessary), while widget updates are done on player
callbacks.

In a few areas in the original code, this pattern was not respected. For
example, the GUI code often used the volume widget to change the
volume: now any instance of changing the volume is done directly through
the player, and the player itself calls a callback that modifies the
volume widget.

For these reasons, the player class now has a bunch of new callbacks.

There were also a few changes regarding style and slight code
semplifications: for example, there are now almost no widgets in the
MainWindow class itself (preferring to just declare them inside the
constructor, the only place where they are used).