Add cache for tracks
Apparently using gme_track_info means I also have to free that info,
which means just repeatedly clicking the next and previous buttons leads
to a fuckload of small allocations, which... Anyway, you get the idea.
In the name of trying not to do too many allocations each milliseconds,
here's a track cache, which is filled any time a file is loaded. (This
changes nothing for files since we still had to do an allocation to get
metadata every time we load a file).