Add interface hierarchy, clean up error handling
I made this hierarchy, written inside format.*, because in the future
I'm going to add support for more formats than those inside GME. It also
nicely divides the GME cruft from other cruft in the Player class.

As a part of the hierarchy, I also had to make an Error struct so that
we may finally get uniform error handling (mostly. There's that play()
function that only gets called in the audio thread that returns
errors...)