Put track info in context menu
I've removed the track info stuff in the second tab and put it a little
'details' option that opens a new window. This was users can even get
info for stuff that is not playing.
Perhaps later I'll see if it's feasible putting this nice stuff on its
own pane without the GUI looking like shit... but I doubt it. (I am much
more in support of a little status above or below the music controls,
and I'll probably implement it soon).